backend/medical-chatbot/requirements.txt
setuptools>=65.0.0
wheel
fastapi>=0.119.0
uvicorn>=0.38.0
python-dotenv>=1.1.0
requests>=2.32.0
faiss-cpu>=1.12.0
sentence-transformers>=5.1.0
pdfplumber>=0.11.0
pypdf>=6.1.0
langchain>=1.0.0
numpy>=2.0.0
torch>=2.9.0
transformers>=4.57.0